Moved ble.ts from BLE folder to src folder
